requirements.txt
numpy~=1.26.1
requests~=2.31.0
Pillow~=10.1.0
pytest~=7.4.2
torch~=2.1.0
matplotlib~=3.8.0
opencv-python~=4.8.1.78
scipy~=1.11.3
imageio~=2.31.5
scikit-image~=0.22.0
setuptools==68.2.2